Rename: 스터디 모집 dto 이동
