+++ src/controller/post-detail.controller.ts
@@ -4,10 +4,10 @@ import { PaginationRequest } from "src/common/pagination/pagination-request";
 import { PaginationResponse } from "src/common/pagination/pagination-response";
 import { ApiPaginatedResponse } from "src/common/pagination/pagination.decorator";
 import { Roles } from "src/common/roles/roles.decorator";
-import { RecruitedPostInfoDto } from "src/dto/request/recruited-post-info";
+import { RecruitedPostInfoDto } from "src/dto/request/post-detail/recruited-post-info";
 import { PostCommentResponse } from "src/dto/response/post-comment.response";
 import { PostDetailResponse } from "src/dto/response/post-detail.response";
-import { RecruitPostResponse } from "src/dto/response/recruit-post-response";
+import { RecruitPostResponse } from "src/dto/response/post-detail/recruit-post-response";
 import { RolesGuard } from "src/guard/roles.guard";
 import { PostDetailService } from "src/service/post-detail.service";
