+++ src/dto/request/post-detail/recruited-post-info.ts
